# openapi.yaml
openapi: 3.1.0
info:
  title: NioDB API
  summary: NioDB — The Agentic DB that works.
  version: 0.1.0
  description: A lightweight database with natural-language queries, powered by Nio.
    Each Rust server has one workspace; clients do not supply workspace IDs. Nio CLI
    powers chat and read queries. AlaSQL executes restricted SQL. Recovery uses a
    checksummed JSONL journal, derived TOON artifacts and content-addressed file blobs.
    Backend credentials provision app users; users log in with username and password.
    Protected routes accept bearer credentials. Use HTTPS for remote access. Responses
    and errors carry X-Request-ID.
servers:
- url: /
  description: Same origin as the deployed NioDB server
security:
- bearerAuth: []
paths:
...
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags:
      - Webhooks
      summary: List webhooks
      description: Backend credential only. Signing secrets are omitted.
      responses:
        '200':
          description: Webhook registrations
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createWebhook
      tags:
      - Webhooks
      summary: Register webhook
      description: Backend credential only. HTTPS URL required; HTTP loopback is allowed. Creation returns the signing secret once. Delivery uses three attempts while the server runs, a five-second timeout and no redirects. The X-NioDB-Signature header is sha256=<hex HMAC-SHA256> of the exact JSON body keyed by the returned secret string. Delivery is not durable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [event_name, url]
              properties:
                event_name:
                  type: string
                  description: Exact event name or '*'.
                url:
                  type: string
                  format: uri
                  maxLength: 2048
      responses:
        '201':
          description: Webhook registered; save secret now
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Webhook'
                - type: object
                  required: [secret]
                  properties:
                    secret:
                      type: string
                      writeOnly: true
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer credential supporting combined client-token:secret-token (or individual client/secret tokens, both with full backend access), or expiring app-user session. Keep backend credentials off public web/mobile clients. Registration requires a backend credential; me/logout require a user session.
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      - request_id
      properties:
        request_id:
          type: string
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              examples:
              - nio_unavailable
              - invalid_request
            message:
              type: string
...
    Webhook:
      type: object
      required: [id, event_name, url, created_at, created_by]
      properties:
        id:
          type: string
        event_name:
          type: string
        url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
  responses:
    Error:
      description: 'Request failed; no credentials, raw CLI output or internal paths.
        Common statuses: 400 invalid request/query, 401 authentication, 403 workspace
        denial, 404 missing record, 409 idempotency/conversation conflict, 413 size
        limit, 422 malformed body, 429 capacity exhausted, 502 invalid Nio output,
        503 dependency/storage unavailable, 504 timeout.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...